requirements.txt
pywin32==302
PyAutoGUI==0.9.53
keyboard==0.13.5
mouse==0.7.1
icrawler==0.6.6
imutils==0.5.4
matplotlib==3.4.3
mss==6.1.0
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
termcolor==1.1.0
torch==1.9.1
win32gui==221.6
